# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Muchi API
  version: 1.0.0
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
  /health:
    get:
      security: []
      responses:
        "200": {description: Healthy}
        "500": {$ref: "#/components/responses/InternalError"}
  /health/sources:
    get:
      responses:
        "200":
          description: Provider health
          content:
            application/json:
              schema:
                type: object
                required: [sources]
                properties:
                  sources: {type: array, items: {$ref: "#/components/schemas/SourceHealth"}}
  /supported-games:
    get:
      responses:
        "200":
          description: Games currently supported by the API
          content:
            application/json:
              schema:
                type: object
                required: [games]
                properties:
                  games:
                    type: array
                    items: {$ref: "#/components/schemas/SupportedGame"}
  /stores/{store_id}/inventory/refresh:
    parameters:
      - name: store_id
        in: path
        required: true
        description: Store identifier or store name, as config/stores.yaml writes it
        schema: {type: string}
    post:
      summary: Forget the cached published inventory of a store
      parameters:
        - name: list
          in: query
          description: Refresh only the list carrying this label
          schema: {type: string}
      responses:
        "200":
          description: Lists forgotten; the next search reads them again
          content:
            application/json:
              schema:
                type: object
                required: [store_id, refreshed]
                properties:
                  store_id: {type: string}
                  list: {type: string}
                  refreshed: {type: integer, minimum: 1}
        "404": {$ref: "#/components/responses/NotFound"}
  /searches:
    post:
      parameters:
        - {$ref: "#/components/parameters/IdempotencyKey"}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateSearch"}
      responses:
        "202":
          description: Search queued
          content: {application/json: {schema: {$ref: "#/components/schemas/Search"}}}
        "409": {$ref: "#/components/responses/Conflict"}
  /searches/{search_id}:
    parameters:
      - {$ref: "#/components/parameters/SearchID"}
    get:
      responses:
        "200":
          description: Search state
          content: {application/json: {schema: {$ref: "#/components/schemas/Search"}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /searches/{search_id}/results:
    parameters:
      - {$ref: "#/components/parameters/SearchID"}
    get:
      parameters:
        - name: after
          in: query
          description: Return results completed after this cursor
          schema: {type: integer, minimum: 0, default: 0}
        - name: limit
          in: query
          description: Maximum results returned per page
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
      responses:
        "200":
          description: Partial or final results
          content: {application/json: {schema: {$ref: "#/components/schemas/SearchResults"}}}
  /searches/{search_id}/stock:
    parameters:
      - {$ref: "#/components/parameters/SearchID"}
    post:
      description: >-
        Visits the stores again for offers this search already found, and
        answers what each one says about its stock right now. A price outlives
        a search; stock does not, so a caller crowning the cheapest offer asks
        here before trusting it. Offers are named by the id the search gave
        them: a url of the caller's choosing is refused, so this is never a
        proxy. Readings come back in the order asked, one per offer, at most
        fifty per request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [offers]
              properties:
                offers:
                  type: array
                  minItems: 1
                  maxItems: 50
                  items: {type: string}
      responses:
        "200":
          description: One stock reading per offer asked
          content:
            application/json:
              schema:
                type: object
                required: [offers]
                properties:
                  offers: {type: array, items: {$ref: "#/components/schemas/OfferStock"}}
        "404": {$ref: "#/components/responses/NotFound"}
  /searches/{search_id}/cancel:
    parameters:
      - {$ref: "#/components/parameters/SearchID"}
      - {$ref: "#/components/parameters/IdempotencyKey"}
    post:
      responses:
        "200":
          description: Search cancelled
          content: {application/json: {schema: {$ref: "#/components/schemas/Search"}}}
        "409": {$ref: "#/components/responses/Conflict"}
  /cards/offers:
    get:
      parameters:
        - name: game
          in: query
          description: Game catalog; defaults to Magic when omitted
          schema:
            allOf: [{$ref: "#/components/schemas/Game"}]
            default: magic
        - name: name
          in: query
          required: true
          schema: {type: string, minLength: 1}
        - name: match
          in: query
          description: How wide a store title may stray from the card name
          schema:
            allOf: [{$ref: "#/components/schemas/MatchMode"}]
            default: exact
        - name: kind
          in: query
          description: Whether to look for single cards or sealed product
          schema:
            allOf: [{$ref: "#/components/schemas/ProductKind"}]
            default: single
      responses:
        "200":
          description: Offers for one card
          content:
            application/json:
              schema:
                type: object
                required: [name, match, kind, offers, faults]
                properties:
                  name: {type: string}
                  match: {$ref: "#/components/schemas/MatchMode"}
                  kind: {$ref: "#/components/schemas/ProductKind"}
                  offers: {type: array, items: {$ref: "#/components/schemas/Offer"}}
                  faults:
                    type: array
                    description: >-
                      Sources that failed while others answered. Empty on a full
                      answer; never absent, so a caller cannot forget to read it.
                      A non-empty list means the offers are incomplete, and the
                      answer is not cached.
                    items: {$ref: "#/components/schemas/SourceFault"}
  /cards/metadata:
    get:
      parameters:
        - name: game
          in: query
          schema:
            allOf: [{$ref: "#/components/schemas/Game"}]
            default: magic
        - {name: name, in: query, required: true, schema: {type: string, minLength: 1}}
        - {name: language, in: query, schema: {type: string}}
        - {name: edition, in: query, schema: {type: string}}
        - {name: foil, in: query, schema: {type: boolean, default: false}}
      responses:
        "200":
          description: Presentation metadata for one card in the selected game
          content: {application/json: {schema: {$ref: "#/components/schemas/CardMetadata"}}}
  /cards/autocomplete:
    get:
      parameters:
        - name: game
          in: query
          schema:
            allOf: [{$ref: "#/components/schemas/Game"}]
            default: magic
        - {name: name, in: query, required: true, schema: {type: string, minLength: 1}}
        - {name: language, in: query, schema: {type: string}}
      responses:
        "200":
          description: Card names matching text in the selected game
          content:
            application/json:
              schema:
                type: object
                required: [suggestions]
                properties:
                  suggestions: {type: array, items: {type: string}}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  parameters:
    SearchID: {name: search_id, in: path, required: true, schema: {type: string}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, maxLength: 128}}
  responses:
    Conflict: {description: Idempotency or state conflict, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
    NotFound: {description: Search not found, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
    InternalError: {description: Internal error, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  schemas:
    CreateSearch:
      type: object
      additionalProperties: false
      required: [cards, options]
      properties:
        game:
          allOf: [{$ref: "#/components/schemas/Game"}]
          description: Game catalog; defaults to Magic when omitted
          default: magic
        cards:
          type: array
          minItems: 1
          maxItems: 500
          items:
            type: object
            additionalProperties: false
            required: [name, quantity]
            properties:
              name: {type: string, minLength: 1}
              quantity: {type: integer, minimum: 1, maximum: 99}
        options:
          type: object
          additionalProperties: false
          required: [verify_stock]
          properties:
            verify_stock: {type: boolean}
            match:
              allOf: [{$ref: "#/components/schemas/MatchMode"}]
              default: exact
            kind:
              allOf: [{$ref: "#/components/schemas/ProductKind"}]
              default: single
    SourceFault:
      type: object
      required: [source, reason]
      properties:
        source: {type: string, description: The source as /v1/health/sources names it}
        reason: {type: string}
    MatchMode:
      type: string
      enum: [exact, includes]
      description: >-
        exact keeps the card and its printings (Kuriboh, Kuriboh (C));
        includes keeps any title carrying the name (Winged Kuriboh, Linkuriboh).
    ProductKind:
      type: string
      enum: [single, sealed]
      description: >-
        single keeps the cards a store sells one by one; sealed keeps the boxes
        it sells unopened (Booster Box, Elite Trainer Box, Bundle, Display).
        A sealed answer groups by the whole product title and its set, because a
        Booster Pack and a Booster Display share every word but one and differ
        by ten times the price.
    Search:
      type: object
      required: [id, status, total, processed, found, not_found, errors, created_at, updated_at]
      properties:
        id: {type: string}
        status: {type: string, enum: [queued, running, completed, completed_with_errors, failed, cancelled]}
        total: {type: integer}
        processed: {type: integer}
        current_card: {type: string}
        found: {type: integer}
        not_found: {type: integer}
        errors: {type: integer}
        created_at: {type: string, format: date-time}
        started_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        finished_at: {type: string, format: date-time}
        incident_id: {type: string}
    SearchResults:
      type: object
      required: [search_id, items, cursor, has_more]
      properties:
        search_id: {type: string}
        items: {type: array, items: {$ref: "#/components/schemas/SearchItem"}}
        cursor: {type: integer, minimum: 0}
        has_more: {type: boolean}
    SearchItem:
      type: object
      required: [id, search_id, game, position, sequence, original_name, normalized_name, quantity, status, attempts, offers]
      properties:
        id: {type: string}
        search_id: {type: string}
        game: {$ref: "#/components/schemas/Game"}
        position: {type: integer}
        sequence:
          type: integer
          minimum: 0
          description: >-
            The order in which the item finished, counting from one. An item
            still pending or running has not earned one and reads zero, so a
            caller sorts the finished ones by this and tells the rest apart by
            status, never by a missing field.
        original_name: {type: string}
        normalized_name: {type: string}
        quantity: {type: integer}
        status: {type: string, enum: [pending, running, found, not_found, source_error]}
        attempts: {type: integer}
        source: {type: string}
        error_code: {type: string}
        error_message: {type: string}
        offers: {type: array, items: {$ref: "#/components/schemas/Offer"}}
    OfferStock:
      type: object
      required: [id, stock_status]
      properties:
        id: {type: string}
        stock_status: {type: string, enum: [available, unavailable, unknown]}
        stock_quantity:
          type: integer
          minimum: 0
          description: >-
            Units the store declares. Absent means the store never said; zero
            means it said none. A store that would not answer reads `unknown`
            and no count, so the caller moves to the next cheapest offer.
    Offer:
      type: object
      required: [id, card_name, store, price_amount, price_currency, url, source, stock_status, suspicious]
      properties:
        id: {type: string}
        card_name: {type: string}
        card_key:
          type: string
          description: >-
            The card this offer is for, with the printing dropped and lowercased.
            `Winged Kuriboh`, `LDS3-EN100 "Winged Kuriboh" Common` and
            `Winged Kuriboh (PUR)` all answer `winged kuriboh`, so one card sold
            by three sources groups as one. Group by this; show `card_name`.
        store: {type: string}
        price_amount: {type: string, pattern: "^[0-9]+(\\.[0-9]{1,2})?$"}
        price_currency: {type: string, pattern: "^[A-Z]{3}$"}
        url: {type: string, format: uri}
        image: {type: string, format: uri}
        variant_id: {type: string}
        language: {type: string}
        condition: {type: string}
        finish: {type: string}
        kind: {$ref: "#/components/schemas/ProductKind"}
        locations:
          type: array
          description: Verified pickup locations, formatted for display.
          items: {type: string}
        source: {type: string}
        stock_status: {type: string, enum: [available, unavailable, unknown]}
        stock_quantity:
          type: integer
          minimum: 0
          description: >-
            Units the store declares. Absent means the store never said; zero
            means it said none. Only some platforms publish a count: shopify
            says whether a variant sells, not how many remain.
        suspicious:
          type: boolean
          description: >-
            Priced far below the median of its peers. Peers are the offers of
            the same card in the same currency, so a wide `includes` search
            never judges one card by another card's prices.
        suspicious_reason: {type: string}
        metadata: {type: object, additionalProperties: {type: string}}
    Game:
      type: string
      enum: [magic, pokemon, yugioh]
      example: magic
    SupportedGame:
      type: object
      required: [name, reference_key]
      properties:
        name: {type: string}
        reference_key: {$ref: "#/components/schemas/Game"}
    CardMetadata:
      type: object
      required: [name]
      properties:
        name: {type: string}
        printed_name: {type: string}
        edition: {type: string}
        image: {type: string, format: uri}
        url: {type: string, format: uri}
    SourceHealth:
      type: object
      required: [source, consecutive_failures, latency_ms]
      properties:
        source: {type: string}
        last_success: {type: string, format: date-time}
        last_failure: {type: string, format: date-time}
        consecutive_failures: {type: integer}
        latency_ms: {type: integer}
        circuit_open_until: {type: string, format: date-time}
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code: {type: string}
        message: {type: string}
        request_id: {type: string}
        incident_id: {type: string}